more changes for 16bit mode number compatibility
I think there will be more, especially inside usermods.

+++ usermods/ST7789_display/ST7789_display.h
@@ -65,7 +65,7 @@ class St7789DisplayUsermod : public Usermod {
     String knownSsid = "";
     IPAddress knownIp;
     uint8_t knownBrightness = 0;
-    uint8_t knownMode = 0;
+    uint16_t knownMode = 0;
     uint8_t knownPalette = 0;
     uint8_t knownEffectSpeed = 0;
     uint8_t knownEffectIntensity = 0;
