+++ wled00/remote.cpp
@@ -119,7 +119,7 @@ static void setOff() {
   }
 }
 
-static void presetWithFallback(uint8_t presetID, uint8_t effectID, uint8_t paletteID) {
+static void presetWithFallback(uint8_t presetID, uint16_t effectID, uint8_t paletteID) {
   resetNightMode();
   unloadPlaylist();
   applyPresetWithFallback(presetID, CALL_MODE_BUTTON_PRESET, effectID, paletteID);
